refactor explcit command folder structure
