+++ services/discord/src/commands/factory/index.ts
